code/analytics/requirements.txt
pandas
scipy
plotly
requests
httpx
scikit-learn
openpyxl
python-dotenv
supabase
psycopg2-binary
anthropic
fastmcp
xdk
kaleido==0.2.1
elevenlabs
matplotlib
Pillow
